changed snackbars where it says 'can not' to 'cannot' for grammar usage
Both the one-word form cannot and the two-word form can not are acceptable, but cannot is more common (in the Oxford English Corpus, three times as common). The two-word form is better only in a construction in which not is part of a set phrase, such as ‘not only … but (also)’: Paul can not only sing well, he also paints brilliantly.